requirements.txt
python-telegram-bot~=21.6
babel~=2.16.0
openpyxl~=3.1.5